+++ import.php
@@ -109,7 +109,7 @@ if ($fromform = $importform->get_data()) {
         throw new moodle_exception('cannotimport', '', $thispageurl->out());
     }
 
-    $result = qbank_importasversion\importer::import_file($qformat, $question, $importfile);
+    $result = qbank_importasversion\importer::import_file($qformat, $question, $importfile, !empty($fromform->force));
 
     // In case anything needs to be done after.
     if (!$qformat->importpostprocess()) {